Extract ReadsSource from DataflowReadsPipeline into a public class in the engine
This class is now needed by other pipelines, so let's make it generally available
in the engine.
